requirements.txt
fastapi
uvicorn
deepface
faiss-cpu
numpy
pillow 